+++ proposals/43:upgrade-11/config.json
@@ -1,5 +1,5 @@
 {
-    "releaseNotes": "https://github.com/Agoric/agoric-sdk/releases/tag/agoric-upgrade-11",
-    "sdkVersion": "36",
-    "planName": "agoric-upgrade-11"
+  "releaseNotes": "https://github.com/Agoric/agoric-sdk/releases/tag/agoric-upgrade-11",
+  "sdkVersion": "36",
+  "planName": "agoric-upgrade-11"
 }
