+++ upgrade-test-scripts/run_actions.sh
@@ -6,9 +6,12 @@ source ./env_setup.sh
 
 export SLOGFILE=slog.slog
 
+PROPOSAL_PATH=$1
+
 startAgd
 
 echo "Agd started. Running actions.sh."
+cd /usr/src/proposals/"$PROPOSAL_PATH/" || exit
 ./actions.sh
 
 echo "Actions completed. Running for a few blocks and exiting."
